app/requirements.txt
# Phoenyra EMS - Core Dependencies
# =================================

# Web Framework
flask>=3.0.0

# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# Optimization (Essential for EMS Intelligence)
cvxpy>=1.4.0
pulp>=2.7.0
scipy>=1.11.0

# Forecasting & Time Series
statsmodels>=0.14.0
prophet>=1.1.0

# HTTP & API Integration
requests>=2.31.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Optional: MQTT for IoT
paho-mqtt>=1.6.0
